+++ docs/CHANGELOG.md
@@ -15,6 +15,10 @@ nav_order: 5
     *Daniel Gonzalez*
     *Reegan Viljoen*
 
+* Fix bug where `#with_request_url`, set the incorrect `request.fullpath`.
+
+    *Nachiket Pusalkar*
+
 ## 3.7.0
 
 * Support Rails 7.1 in CI.